added new file js + refactor const and array
