+++ scripts/data-card.js
@@ -1,0 +1,26 @@
+export const initialCards = [
+  {
+    name: 'Санкт-Петербург',
+    link: 'https://images.unsplash.com/photo-1642250399033-ba2a24390f8c?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1887&q=80'
+  },
+  {
+    name: 'Мурадым, Башкортостан',
+    link: 'https://images.unsplash.com/photo-1604250016587-d6e27c900af2?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1174&q=80'
+  },
+  {
+    name: 'Москва',
+    link: 'https://images.unsplash.com/photo-1498582475317-a8e7fb3e4a68?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1664&q=80'
+  },
+  {
+    name: 'Красноярск',
+    link: 'https://images.unsplash.com/photo-1454614146802-84004a53428d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1074&q=80'
+  },
+  {
+    name: 'Дивногорск',
+    link: 'https://images.unsplash.com/photo-1579615253458-4c740233a960?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1887&q=80'
+  },
+  {
+    name: 'Байкал',
+    link: 'https://images.unsplash.com/photo-1552735855-557bdba3961a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1889&q=80'
+  }
+];
