+++ scripts/data-validation.js
@@ -1,0 +1,8 @@
+export const validationConfig = {
+  formSelector: '.form',
+  inputSelector: '.form__input',
+  submitButtonSelector: '.button_type_save',
+  inactiveButtonClass: 'button_type_disable',
+  inputErrorClass: 'form__input-error_active',
+  errorClass: 'form__input-error',
+};
