--- scripts/elements.js
@@ -1,60 +1,0 @@
-//POPUP
-const popups = document.querySelectorAll('.popup');
-const popupCard = document.querySelector('.popup_add-card');
-const popupProfile = document.querySelector('.popup_edit-profile');
-const popupImg = document.querySelector('.popup_image');
-//PROFILE
-const formProfile = document.querySelector('[name="formProfile"]');
-const profileName = document.querySelector('.profile__name');
-const formInputName = document.querySelector('.form__input_text_name');
-const profileAboutUser = document.querySelector('.profile__info');
-const formAboutUser = document.querySelector('.form__input_text_about');
-//CARD
-const cardTemplate = document.querySelector('#card-template').content;
-const cardContainer = document.querySelector('.cards__list');
-const cardImg = document.querySelector('.cards__image');
-const formCard = document.querySelector('[name="formCard"]');
-const cardInputName = document.querySelector('[name="nameCard"]');
-const cardImgLink = document.querySelector('[name="linkCard"]');
-const imgFigure = document.querySelector('.figure__image');
-const infoFigure = document.querySelector('.figure__info');
-//BUTTON
-const openPopupBtn = document.querySelector('.button_type_edit');
-const addPopupBtn = document.querySelector('.button_type_add');
-const saveCardBtn = formCard.querySelector('.button_type_save');
-
-const validationConfig = {
-  formSelector: '.form',
-  inputSelector: '.form__input',
-  submitButtonSelector: '.button_type_save',
-  inactiveButtonClass: 'button_type_disable',
-  inputErrorClass: 'form__input-error_active',
-  errorClass: 'form__input-error',
-};
-
-const initialCards = [
-  {
-    name: 'Санкт-Петербург',
-    link: 'https://images.unsplash.com/photo-1642250399033-ba2a24390f8c?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1887&q=80'
-  },
-  {
-    name: 'Мурадым, Башкортостан',
-    link: 'https://images.unsplash.com/photo-1604250016587-d6e27c900af2?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1174&q=80'
-  },
-  {
-    name: 'Москва',
-    link: 'https://images.unsplash.com/photo-1498582475317-a8e7fb3e4a68?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1664&q=80'
-  },
-  {
-    name: 'Красноярск',
-    link: 'https://images.unsplash.com/photo-1454614146802-84004a53428d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1074&q=80'
-  },
-  {
-    name: 'Дивногорск',
-    link: 'https://images.unsplash.com/photo-1579615253458-4c740233a960?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1887&q=80'
-  },
-  {
-    name: 'Байкал',
-    link: 'https://images.unsplash.com/photo-1552735855-557bdba3961a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1889&q=80'
-  }
-];
